refactor(backend): extract metrics and client accessor
extract metrics cache and shared client accessor from GuildService for
reusability. phase 0: discordClientAccessor.ts. phase 1: MetricsCache.ts.
no behavior change.
