+++ packages/backend/src/services/GuildService.ts
@@ -1,29 +1,18 @@
 import { z } from 'zod'
 import type { Client, Guild } from 'discord.js'
 import { discordOAuthService, type DiscordGuild } from './DiscordOAuthService'
+import {
+    setClient as setDiscordClient,
+    getClient as getDiscordClient,
+    getServableGuild,
+} from '../utils/discordClientAccessor'
+import { metricsService, type GuildMetrics } from './MetricsCache'
 import { debugLog, errorLog } from '@lucky/shared/utils'
 
-let botClient: Client | null = null
 const DISCORD_API_BASE_URL = 'https://discord.com/api/v10'
 const BOT_GUILD_CACHE_TTL_MS = 60_000
-const GUILD_METRICS_CACHE_TTL_MS = 30_000
 // Discord permission bitfield literal representing zero permissions granted
 const NO_PERMISSIONS = '0'
-
-interface GuildMetrics {
-    memberCount: number | null
-    categoryCount: number | null
-    textChannelCount: number | null
-    voiceChannelCount: number | null
-    roleCount: number | null
-}
-
-interface DiscordGuildChannel {
-    id?: string
-    name?: string
-    type: number
-    position?: number
-}
 
 interface DiscordGuildRole {
     id: string
@@ -79,13 +68,6 @@ export interface GuildEmojiOption {
 }
 
 // Zod validation schemas for Discord API responses
-const discordGuildChannelSchema = z.object({
-    id: z.string().optional(),
-    name: z.string().optional(),
-    type: z.number().int(),
-    position: z.number().optional(),
-})
-
 const discordGuildRoleSchema = z.object({
     id: z.string().min(1),
     name: z.string(),
@@ -98,12 +80,8 @@ const discordGuildRoleSchema = z.object({
 })
 
 export function setBotClient(client: Client | null): void {
-    botClient = client
+    setDiscordClient(client)
     guildService.clearBotGuildCache()
-}
-
-function getClient(): Client | null {
-    return botClient
 }
 
 export interface GuildWithBotStatus extends DiscordGuild {
@@ -122,34 +100,16 @@ class GuildService {
         expiresAt: number
     } | null = null
 
-    private guildMetricsCache = new Map<
-        string,
-        { data: GuildMetrics; expiresAt: number }
-    >()
-
     private botGuildIdsInFlight: Promise<Set<string> | null> | null = null
 
     private getBotClient(): Client | null {
-        return getClient()
-    }
-
-    private async getServableGuild(guildId: string): Promise<Guild | null> {
-        const client = this.getBotClient()
-        if (!client) return null
-        try {
-            return (
-                client.guilds.cache.get(guildId) ??
-                (await client.guilds.fetch(guildId))
-            )
-        } catch {
-            return null
-        }
+        return getDiscordClient()
     }
 
     clearBotGuildCache(): void {
         this.botGuildIdsCache = null
         this.botGuildIdsInFlight = null
-        this.guildMetricsCache.clear()
+        metricsService.clearCache()
     }
 
     private getBotToken(): string | null {
@@ -157,7 +117,12 @@ class GuildService {
         return token && token.length > 0 ? token : null
     }
 
-    private validateChannelArray(data: unknown): DiscordGuildChannel[] {
+    private validateChannelArray(data: unknown): Array<{
+        id?: string
+        name?: string
+        type: number
+        position?: number
+    }> {
         if (!Array.isArray(data)) {
             errorLog({
                 message: 'Invalid channels response from Discord API',
@@ -166,9 +131,21 @@ class GuildService {
             return []
         }
 
-        const validated: DiscordGuildChannel[] = []
+        const channelSchema = z.object({
+            id: z.string().optional(),
+            name: z.string().optional(),
+            type: z.number().int(),
+            position: z.number().optional(),
+        })
+
+        const validated: Array<{
+            id?: string
+            name?: string
+            type: number
+            position?: number
+        }> = []
         for (const item of data) {
-            const result = discordGuildChannelSchema.safeParse(item)
+            const result = channelSchema.safeParse(item)
             if (result.success) {
                 validated.push(result.data)
             } else {
@@ -303,245 +280,12 @@ class GuildService {
         return this.botGuildIdsInFlight
     }
 
-    private getCachedMetrics(guildId: string): GuildMetrics | null {
-        const cached = this.guildMetricsCache.get(guildId)
-        if (!cached) {
-            return null
-        }
-
-        if (cached.expiresAt <= Date.now()) {
-            this.guildMetricsCache.delete(guildId)
-            return null
-        }
-
-        return cached.data
-    }
-
-    private setCachedMetrics(guildId: string, metrics: GuildMetrics): void {
-        this.guildMetricsCache.set(guildId, {
-            data: metrics,
-            expiresAt: Date.now() + GUILD_METRICS_CACHE_TTL_MS,
-        })
-    }
-
-    private emptyMetrics(): GuildMetrics {
-        return {
-            memberCount: null,
-            categoryCount: null,
-            textChannelCount: null,
-            voiceChannelCount: null,
-            roleCount: null,
-        }
-    }
-
-    private countChannelTypes(
-        channels: DiscordGuildChannel[],
-    ): Pick<
-        GuildMetrics,
-        'categoryCount' | 'textChannelCount' | 'voiceChannelCount'
-    > {
-        let categoryCount = 0
-        let textChannelCount = 0
-        let voiceChannelCount = 0
-
-        for (const channel of channels) {
-            if (channel.type === 4) {
-                categoryCount += 1
-                continue
-            }
-
-            if (channel.type === 2 || channel.type === 13) {
-                voiceChannelCount += 1
-                continue
-            }
-
-            if (
-                channel.type === 0 ||
-                channel.type === 5 ||
-                channel.type === 15 ||
-                channel.type === 16
-            ) {
-                textChannelCount += 1
-            }
-        }
-
-        return {
-            categoryCount,
-            textChannelCount,
-            voiceChannelCount,
-        }
-    }
-
-    private mergeMetrics(
-        primary: GuildMetrics,
-        fallback: GuildMetrics,
-    ): GuildMetrics {
-        return {
-            memberCount: primary.memberCount ?? fallback.memberCount,
-            categoryCount: primary.categoryCount ?? fallback.categoryCount,
-            textChannelCount:
-                primary.textChannelCount ?? fallback.textChannelCount,
-            voiceChannelCount:
-                primary.voiceChannelCount ?? fallback.voiceChannelCount,
-            roleCount: primary.roleCount ?? fallback.roleCount,
-        }
-    }
-
-    private buildMetricsFromClientGuild(guild: Guild): GuildMetrics {
-        const channelsCache = guild.channels?.cache
-        const channels =
-            channelsCache && channelsCache.size > 0
-                ? [...channelsCache.values()].map((channel) => ({
-                      type: channel.type,
-                  }))
-                : []
-        const hasChannelsSnapshot = channels.length > 0
-        const counts = hasChannelsSnapshot
-            ? this.countChannelTypes(channels)
-            : null
-        const roleCache = guild.roles?.cache
-
-        return {
-            memberCount: guild.memberCount || null,
-            categoryCount: counts?.categoryCount ?? null,
-            textChannelCount: counts?.textChannelCount ?? null,
-            voiceChannelCount: counts?.voiceChannelCount ?? null,
-            roleCount: roleCache ? roleCache.size || null : null,
-        }
-    }
-
-    private hasUnknownMetrics(metrics: GuildMetrics): boolean {
-        return (
-            metrics.memberCount === null ||
-            metrics.categoryCount === null ||
-            metrics.textChannelCount === null ||
-            metrics.voiceChannelCount === null ||
-            metrics.roleCount === null
-        )
-    }
-
-    private async resolveClientMetrics(
-        guildId: string,
-    ): Promise<GuildMetrics | null> {
-        const client = this.getBotClient()
-        if (!client) {
-            return null
-        }
-
-        const guild = client.guilds.cache.get(guildId)
-        if (!guild) {
-            return null
-        }
-
-        const metricsFromClient = this.buildMetricsFromClientGuild(guild)
-        if (!this.hasUnknownMetrics(metricsFromClient)) {
-            return metricsFromClient
-        }
-
-        const metricsFromApi = await this.fetchGuildMetricsFromApi(guildId)
-        return this.mergeMetrics(metricsFromClient, metricsFromApi)
-    }
-
-    private async fetchGuildMetricsFromApi(
-        guildId: string,
-    ): Promise<GuildMetrics> {
-        const token = this.getBotToken()
-        if (!token) {
-            return this.emptyMetrics()
-        }
-
-        try {
-            const headers = {
-                Authorization: `Bot ${token}`,
-            }
-
-            const [guildResponse, channelsResponse, rolesResponse] =
-                await Promise.all([
-                    fetch(
-                        `${DISCORD_API_BASE_URL}/guilds/${guildId}?with_counts=true`,
-                        { headers, signal: AbortSignal.timeout(10_000) },
-                    ),
-                    fetch(
-                        `${DISCORD_API_BASE_URL}/guilds/${guildId}/channels`,
-                        {
-                            headers,
-                            signal: AbortSignal.timeout(10_000),
-                        },
-                    ),
-                    fetch(`${DISCORD_API_BASE_URL}/guilds/${guildId}/roles`, {
-                        headers,
-                        signal: AbortSignal.timeout(10_000),
-                    }),
-                ])
-
-            if (!guildResponse.ok) {
-                return this.emptyMetrics()
-            }
-
-            const guildPayload = (await guildResponse.json()) as {
-                approximate_member_count?: number
-                member_count?: number
-            }
-
-            const channelsPayload = channelsResponse.ok
-                ? this.validateChannelArray(await channelsResponse.json())
-                : []
-
-            // Only the array length is used here (not individual role
-            // fields), so a shape check is enough — validateRoleArray's full
-            // per-item schema would silently drop roles missing optional
-            // fields this endpoint never reads, undercounting them.
-            const rawRolesPayload = rolesResponse.ok
-                ? await rolesResponse.json()
-                : []
-            const roleCount = Array.isArray(rawRolesPayload)
-                ? rawRolesPayload.length
-                : null
-
-            const counts = this.countChannelTypes(channelsPayload)
-
-            return {
-                memberCount:
-                    guildPayload.approximate_member_count ??
-                    guildPayload.member_count ??
-                    null,
-                categoryCount: counts.categoryCount,
-                textChannelCount: counts.textChannelCount,
-                voiceChannelCount: counts.voiceChannelCount,
-                roleCount: roleCount || null,
-            }
-        } catch (error) {
-            errorLog({
-                message: 'Error fetching guild metrics from Discord API',
-                error,
-            })
-            return this.emptyMetrics()
-        }
-    }
-
     async hasBotInGuild(guildId: string): Promise<boolean> {
         const botGuildIds = await this.getBotGuildIds()
         return (
             this.checkBotInGuild(guildId) ||
             (botGuildIds?.has(guildId) ?? false)
         )
-    }
-
-    async getGuildMetrics(guildId: string): Promise<GuildMetrics> {
-        const cached = this.getCachedMetrics(guildId)
-        if (cached) {
-            return cached
-        }
-
-        const clientMetrics = await this.resolveClientMetrics(guildId)
-        if (clientMetrics) {
-            this.setCachedMetrics(guildId, clientMetrics)
-            return clientMetrics
-        }
-
-        const metrics = await this.fetchGuildMetricsFromApi(guildId)
-        this.setCachedMetrics(guildId, metrics)
-        return metrics
     }
 
     async getGuildMemberContext(
@@ -730,26 +474,51 @@ class GuildService {
 
             const payload = this.validateChannelArray(await response.json())
 
-            return payload
-                .filter(
+            const filteredChannels = payload.filter(
+                (
+                    channel,
+                ): channel is {
+                    id: string
+                    name: string
+                    type: number
+                    position?: number
+                } =>
+                    typeof channel.id === 'string' &&
+                    typeof channel.name === 'string' &&
+                    (channel.type === 0 ||
+                        channel.type === 5 ||
+                        channel.type === 15 ||
+                        channel.type === 16),
+            )
+
+            return filteredChannels
+                .sort(
                     (
-                        channel,
-                    ): channel is typeof channel & {
+                        a: {
+                            id: string
+                            name: string
+                            type: number
+                            position?: number
+                        },
+                        b: {
+                            id: string
+                            name: string
+                            type: number
+                            position?: number
+                        },
+                    ) => (a.position ?? 0) - (b.position ?? 0),
+                )
+                .map(
+                    (channel: {
                         id: string
                         name: string
-                    } =>
-                        typeof channel.id === 'string' &&
-                        typeof channel.name === 'string' &&
-                        (channel.type === 0 ||
-                            channel.type === 5 ||
-                            channel.type === 15 ||
-                            channel.type === 16),
+                        type: number
+                        position?: number
+                    }) => ({
+                        id: channel.id,
+                        name: `#${channel.name}`,
+                    }),
                 )
-                .sort((a, b) => (a.position ?? 0) - (b.position ?? 0))
-                .map((channel) => ({
-                    id: channel.id,
-                    name: `#${channel.name}`,
-                }))
         } catch (error) {
             errorLog({
                 message: 'Failed to fetch guild channels',
@@ -764,7 +533,7 @@ class GuildService {
 
         if (client) {
             try {
-                const guild = await this.getServableGuild(guildId)
+                const guild = await getServableGuild(guildId)
                 if (guild) {
                     return [...guild.emojis.cache.values()].map((emoji) => ({
                         id: emoji.id,
@@ -898,8 +667,14 @@ class GuildService {
                     ? undefined
                     : this.generateBotInviteUrl(guild.id)
                 const metrics = hasBot
-                    ? await this.getGuildMetrics(guild.id)
-                    : this.emptyMetrics()
+                    ? await metricsService.getGuildMetrics(guild.id)
+                    : {
+                          memberCount: null,
+                          categoryCount: null,
+                          textChannelCount: null,
+                          voiceChannelCount: null,
+                          roleCount: null,
+                      }
 
                 return {
                     ...guild,
@@ -926,7 +701,7 @@ class GuildService {
 
         const hasBot = true
         const botInviteUrl = this.generateBotInviteUrl(guildId)
-        const metrics = await this.getGuildMetrics(guildId)
+        const metrics = await metricsService.getGuildMetrics(guildId)
 
         return {
             id: guild.id,
@@ -961,7 +736,7 @@ class GuildService {
 
         return Promise.all(
             guilds.map(async (guild) => {
-                const metrics = await this.getGuildMetrics(guild.id)
+                const metrics = await metricsService.getGuildMetrics(guild.id)
                 const iconUrl = guild.icon
                     ? `https://cdn.discordapp.com/icons/${guild.id}/${guild.icon}.png?size=64`
                     : null
@@ -1058,7 +833,7 @@ class GuildService {
         guildId: string,
         data: RoleUpsertData,
     ): Promise<GuildRoleManage> {
-        const guild = await this.getServableGuild(guildId)
+        const guild = await getServableGuild(guildId)
 
         if (guild) {
             const role = await guild.roles.create({
@@ -1145,7 +920,7 @@ class GuildService {
         roleId: string,
         data: RoleUpsertData,
     ): Promise<GuildRoleManage> {
-        const guild = await this.getServableGuild(guildId)
+        const guild = await getServableGuild(guildId)
 
         if (guild) {
             const role = await guild.roles.fetch(roleId)
@@ -1231,7 +1006,7 @@ class GuildService {
     }
 
     async deleteGuildRole(guildId: string, roleId: string): Promise<void> {
-        const guild = await this.getServableGuild(guildId)
+        const guild = await getServableGuild(guildId)
 
         if (guild) {
             const role = await guild.roles.fetch(roleId)
